requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
#	 CDDA 10.2 + PyTorch 1.8
#    pip install torch==1.8.1+cu102 torchvision==0.9.1+cu102 torchaudio===0.8.1 -f https://download.pytorch.org/whl/torch_stable.html
#    pip install pytorch-lightning
#    pip install pytorch-lightning-bolts
#    pip install wandb --user
#    pip install gym --user

tensorboard==2.5.0                     
tensorboard-data-server==0.6.0                     
tensorboard-plugin-wit==1.8.0                     

tensorflow-estimator==1.15.1                    
tensorflow-gpu==2.2.0                     
tensorflow-gpu-estimator==2.2.0                     

torchsummary==1.5.1                     
pytorch-lightning==1.2.6                     
pytorch-lightning-bolts==0.3.0                     


torch==1.8.1+cu102               
torch==1.8.1                     
tqdm==4.41.0                    
PyYAML==5.3.1 
numpy==1.19.5                       

#    For Albumentations
opencv-python==4.1.1.26                  
opencv-python-headless==4.4.0.46                  
albumentations==0.5.2     
Pillow==7.2.0        
dlib==19.21.1    
